Add TextStorageCore
`TextStorage` manages both the abstraction for `NSTextStorage` and provides the storage management for the text itself. In this commit we introduce a new class, `TextStorageCore` which manages the storage. This class is open, allowing for customization. Some customizations in mind include "temporal" attributes which are not persisted but computed at display time (for example, detecting and rendering phone numbers in a different manner, but not storing the rendering in the underlying storage).

To do this, one can inherit `TextStorageCore` and add another `NSAttributedString`, and then override the necessary methods to "merge" attributes as needed.